// apps/web/src/lib/service-clients/service-cognition/generated/schemas/extractInputRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Document Cognition Service
 * OpenAPI spec version: 1.0.0
 */
import type { InputIntent } from './inputIntent';

export interface ExtractInputRequest {
  intent: InputIntent;
  reference_time: string;
  /** @minimum 0 */
  revision: number;
  text: string;
  time_zone: string;
}